# [P0] ARCH-001: Complete System Architecture Documentation & Security Implementation