--- a/voice-terminal-components/voice-terminal-components/src/components/HexagonGrid/HexagonGrid.stories.tsx
+++ b/voice-terminal-components/voice-terminal-components/src/components/HexagonGrid/HexagonGrid.stories.tsx
@@ -92,6 +92,33 @@ It starts with a single hexagon at rest and expands to multiple rings based on a
       },
       description: 'Total size of the grid in pixels (100-250px recommended)',
     },
+    blurIntensity: {
+      control: { 
+        type: 'range', 
+        min: 0.5, 
+        max: 2.0, 
+        step: 0.1 
+      },
+      description: 'Shadow blur intensity (0.5 = sharper, 2.0 = softer)',
+    },
+    strokeIntensity: {
+      control: { 
+        type: 'range', 
+        min: 0, 
+        max: 2.0, 
+        step: 0.1 
+      },
+      description: 'Stroke intensity (0 = no strokes, 1 = normal, 2 = strong)',
+    },
+    gradientIntensity: {
+      control: { 
+        type: 'range', 
+        min: 0, 
+        max: 2.0, 
+        step: 0.1 
+      },
+      description: 'Gradient intensity (0 = no gradients, 1 = normal, 2 = strong)',
+    },
     className: {
       control: 'text',
       description: 'Additional CSS class names',
@@ -449,6 +476,100 @@ export const LargeSize: Story = {
     docs: {
       description: {
         story: 'Large 250px grid - ideal for prominent voice indicators or hero sections.',
+      },
+    },
+  },
+};
+
+// Intensity variations
+export const IntensityControls: Story = {
+  args: {
+    amplitude: 50,
+    size: 200,
+    blurIntensity: 1.2,
+    strokeIntensity: 1,
+    gradientIntensity: 1,
+  },
+  parameters: {
+    docs: {
+      description: {
+        story: `
+Test different intensity settings for shadows, strokes, and gradients:
+
+- **Blur Intensity**: Controls shadow softness (0.5 = sharp, 2.0 = very soft)
+- **Stroke Intensity**: Controls edge definition (0 = no edges, 2 = strong edges)  
+- **Gradient Intensity**: Controls internal shading (0 = flat, 2 = dramatic gradients)
+
+Try different combinations to achieve your desired visual effect.
+        `,
+      },
+    },
+  },
+};
+
+export const SharpShadows: Story = {
+  args: {
+    amplitude: 50,
+    size: 200,
+    blurIntensity: 0.5,
+    strokeIntensity: 1.5,
+    gradientIntensity: 1,
+  },
+  parameters: {
+    docs: {
+      description: {
+        story: 'Sharp shadows with enhanced strokes for crisp paper-cut appearance.',
+      },
+    },
+  },
+};
+
+export const SoftShadows: Story = {
+  args: {
+    amplitude: 50,
+    size: 200,
+    blurIntensity: 2.0,
+    strokeIntensity: 0.5,
+    gradientIntensity: 1,
+  },
+  parameters: {
+    docs: {
+      description: {
+        story: 'Soft, diffused shadows with subtle strokes for dreamy effect.',
+      },
+    },
+  },
+};
+
+export const DramaticGradients: Story = {
+  args: {
+    amplitude: 50,
+    size: 200,
+    blurIntensity: 1.2,
+    strokeIntensity: 1,
+    gradientIntensity: 2.0,
+  },
+  parameters: {
+    docs: {
+      description: {
+        story: 'Enhanced gradient intensity for more dramatic internal shading.',
+      },
+    },
+  },
+};
+
+export const FlatAppearance: Story = {
+  args: {
+    amplitude: 50,
+    size: 200,
+    blurIntensity: 1.0,
+    strokeIntensity: 0,
+    gradientIntensity: 0,
+  },
+  parameters: {
+    docs: {
+      description: {
+        story: 'Flat appearance with no strokes or gradients, shadows only.',
       },
     },
   },