--- a/voice-terminal-hybrid/.storybook/.gitignore
+++ b/voice-terminal-hybrid/.storybook/.gitignore
@@ -1,0 +1,19 @@
+# Storybook build outputs
+storybook-static/
+
+# Temporary files
+*.tmp
+*.temp
+
+# Logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# Coverage directory used by tools like istanbul
+coverage/
+
+# Cache
+.cache/
+node_modules/.cache/