… PARALLEL_DEVELOPMENT_GUIDE.md
… architecture/ADVANCED_RBAC_SYSTEM.md
… architecture/API_RATE_LIMITING_SYSTEM.md
… architecture/API_SECURITY_FRAMEWORK.md
… architecture/BACKEND_SECURITY_FRAMEWORK.md
… architecture/BACKPRESSURE_MANAGEMENT.md
… architecture/BACKUP_DISASTER_RECOVERY.md
… architecture/CACHING_STRATEGY.md
… architecture/CIRCUIT_BREAKER_IMPLEMENTATION.md
… architecture/COMMAND_SAFETY_ENGINE.md
… architecture/CONNECTION_POOL_MANAGEMENT.md
… architecture/DATABASE_SECURITY_HARDENING.md
… architecture/DATA_FLOW_ARCHITECTURE.md
… architecture/EMERGENCY_SECURITY_MEASURES.md
… architecture/ENCRYPTION_KEY_MANAGEMENT.md
… architecture/IPC_SECURITY_HARDENING.md
… architecture/MFA_AUTHENTICATION_SYSTEM.md
… architecture/MICROSERVICE_SECURITY_ARCHITECTURE.md
… architecture/PERFORMANCE_MONITORING_SECURITY.md
… architecture/PERFORMANCE_MONITORING_TUNING.md
… architecture/PHASE_0_IMPLEMENTATION_SUMMARY.md
… architecture/PHASE_1_IMPLEMENTATION_SUMMARY.md
… architecture/PRODUCTION_INPUT_VALIDATION.md
… architecture/README.md
… architecture/SECURITY_COMPLIANCE_FRAMEWORK.md
… architecture/SECURITY_CONFIGURATION_TEMPLATES.md
… architecture/SECURITY_MONITORING_SIEM.md
… architecture/SECURITY_REVIEW_ASSESSMENT.md
… architecture/SESSION_MANAGEMENT_INFRASTRUCTURE.md
… architecture/STREAM_PROCESSING_OPTIMIZATION.md